% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{addOntologyTerms}
\alias{addOntologyTerms}
\title{Add annotations to the ontology tree}
\usage{
addOntologyTerms(tree, annotations)
}
\arguments{
\item{tree}{An OntologyTree object}

\item{annotations}{Term annotation list}
}
\value{
Updated OntologyTree object
}
\description{
Add annotations to the ontology tree
}
\author{
Kai Guo
}
