% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{addSuperRoot}
\alias{addSuperRoot}
\title{Add super root node to handle multiple roots}
\usage{
addSuperRoot(roots, parentMap, childMap, relationMap, terms, verbose)
}
\arguments{
\item{roots}{Root term indices}

\item{parentMap}{Parent mapping}

\item{childMap}{Child mapping}

\item{relationMap}{Relation mapping}

\item{terms}{Term names}

\item{verbose}{Print messages flag}
}
\value{
List of updated mappings
}
\description{
Add super root node to handle multiple roots
}
\author{
Kai Guo
}
\keyword{internal}
