% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{addTermStats}
\alias{addTermStats}
\title{Add term statistics to OntologyTree}
\usage{
addTermStats(tree, name, data)
}
\arguments{
\item{tree}{OntologyTree object}

\item{name}{Name of the statistic}

\item{data}{Data to store (must have same length as number of terms)}
}
\value{
Updated OntologyTree object
}
\description{
Add term statistics to OntologyTree
}
\author{
Kai Guo
}
