% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{buildChildMap}
\alias{buildChildMap}
\title{Build child term mapping}
\usage{
buildChildMap(parents, children, termIndices, termCount)
}
\arguments{
\item{parents}{Parent terms}

\item{children}{Child terms}

\item{termIndices}{Term index mapping}

\item{termCount}{Total term count}
}
\value{
Child mapping list
}
\description{
Build child term mapping
}
\author{
Kai Guo
}
\keyword{internal}
