% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{buildGOTree}
\alias{buildGOTree}
\title{Build Gene Ontology Tree}
\usage{
buildGOTree(
  namespace = c("BP", "CC", "MF"),
  keytype = "SYMBOL",
  relations = "part_of",
  orgDb = NULL,
  evidenceCode = NULL,
  includeAlt = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{namespace}{GO namespace ('BP', 'CC', or 'MF')}

\item{relations}{Character vector of relation types to include}

\item{orgDb}{Organism annotation database name or object}

\item{evidenceCode}{Vector of evidence codes to filter}

\item{includeAlt}{Include alternative terms}

\item{verbose}{Print progress messages}
}
\value{
OntologyTree object
}
\description{
Build Gene Ontology Tree
}
\examples{
# Basic BP ontology
go_bp <- buildGOTree("BP")

# MF ontology with specific relations
go_mf <- buildGOTree("MF", relations = c("part_of", "regulates"))

# With organism annotations
\dontrun{
tree <- buildGOTree("BP", orgDb = "org.Hs.eg.db")
}
}
\author{
Kai Guo
}
