% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{buildOntologyTree}
\alias{buildOntologyTree}
\title{Create an ontology tree}
\usage{
buildOntologyTree(
  parentTerms,
  childTerms = NULL,
  edgeTypes = NULL,
  relTree = NULL,
  source = "Ontology",
  annotations = NULL,
  removeCycles = FALSE,
  removeIsolated = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{parentTerms}{Character vector of parent terms}

\item{childTerms}{Character vector of child terms}

\item{edgeTypes}{Character vector of relation types}

\item{relTree}{Ontology tree object for relation hierarchy}

\item{source}{Data source identifier}

\item{annotations}{Term annotations list}

\item{removeCycles}{Remove cyclic paths if TRUE}

\item{removeIsolated}{Remove isolated rings}

\item{verbose}{Print progress messages}
}
\value{
OntologyTree object
}
\description{
Create an ontology tree
}
\author{
Kai Guo
}
