% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{buildParentMap}
\alias{buildParentMap}
\title{Build parent term mapping}
\usage{
buildParentMap(parents, children, termIndices, termCount)
}
\arguments{
\item{parents}{Parent terms}

\item{children}{Child terms}

\item{termIndices}{Term index mapping}

\item{termCount}{Total term count}
}
\value{
Parent mapping list
}
\description{
Build parent term mapping
}
\author{
Kai Guo
}
\keyword{internal}
