% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{buildRelationMap}
\alias{buildRelationMap}
\title{Build relation type mapping}
\usage{
buildRelationMap(relations, parents, termIndices, termCount)
}
\arguments{
\item{relations}{Relation types}

\item{parents}{Parent terms}

\item{termIndices}{Term index mapping}

\item{termCount}{Total term count}
}
\value{
Relation mapping list
}
\description{
Build relation type mapping
}
\author{
Kai Guo
}
\keyword{internal}
