% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{buildTreeStructure}
\alias{buildTreeStructure}
\title{Build the core tree structure}
\usage{
buildTreeStructure(df, annotations, namespace, includeAlt, verbose)
}
\description{
Build the core tree structure
}
\author{
Kai Guo
}
\keyword{internal}
