% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateAnnotationIC}
\alias{calculateAnnotationIC}
\title{Calculate annotation-based Information Content}
\usage{
calculateAnnotationIC(tree, uniquify = TRUE, useCache = TRUE, verbose = TRUE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{uniquify}{Whether to use unique annotations}

\item{useCache}{Whether to use cached values}

\item{verbose}{Whether to print messages}
}
\value{
Numeric vector of IC values
}
\description{
Calculate annotation-based Information Content
}
\author{
Kai Guo
}
