% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateIC}
\alias{calculateIC}
\title{Calculate Information Content for ontology terms}
\usage{
calculateIC(
  tree,
  method = c("universal", "annotation", "wang", "offspring"),
  useCache = TRUE,
  contributionFactor = c(is_a = 0.8, part_of = 0.6),
  uniquify = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{tree}{OntologyTree object}

\item{method}{IC calculation method ("universal", "annotation","wang" or "offspring")}

\item{useCache}{Whether to use cached values}

\item{contributionFactor}{Named numeric vector of semantic contribution factors}

\item{uniquify}{Whether to use unique annotations (annotation method)}

\item{verbose}{Whether to print messages}
}
\value{
Named numeric vector of IC values
}
\description{
Calculate Information Content for ontology terms
}
\author{
Kai Guo
}
