% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{calculateKappaSimilarity}
\alias{calculateKappaSimilarity}
\title{Calculate Kappa similarity between terms}
\usage{
calculateKappaSimilarity(annotMatrix)
}
\arguments{
\item{annotMatrix}{Annotation matrix}
}
\value{
Similarity matrix
}
\description{
Calculate Kappa similarity between terms
}
\author{
Kai Guo
}
\keyword{internal}
