% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{calculateOverlapSimilarity}
\alias{calculateOverlapSimilarity}
\title{Calculate overlap similarity between terms}
\usage{
calculateOverlapSimilarity(annotMatrix)
}
\arguments{
\item{annotMatrix}{Annotation matrix}
}
\value{
Similarity matrix
}
\description{
Calculate overlap similarity between terms
}
\author{
Kai Guo
}
\keyword{internal}
