% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateWangIC}
\alias{calculateWangIC}
\title{Calculate Wang's Information Content}
\usage{
calculateWangIC(
  tree,
  contributionFactor = c(is_a = 0.8, part_of = 0.6),
  useCache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{tree}{OntologyTree object}

\item{contributionFactor}{Named numeric vector of semantic contribution factors}

\item{useCache}{Whether to use cached values}

\item{verbose}{Whether to print messages}
}
\value{
Numeric vector of IC values
}
\description{
Calculate Wang's Information Content
}
\author{
Kai Guo
}
