% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertToIgraph}
\alias{convertToIgraph}
\title{Convert OntologyTree to igraph object}
\usage{
convertToIgraph(tree)
}
\arguments{
\item{tree}{An OntologyTree object to convert}
}
\value{
An igraph object representing the ontology structure
}
\description{
Convert OntologyTree to igraph object
}
\examples{
\dontrun{
g <- convertToIgraph(tree)
}
}
\author{
Kai Guo
}
