% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{cross_multiply}
\alias{cross_multiply}
\title{Cross multiplication helper}
\usage{
cross_multiply(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Cross product matrix
}
\description{
Cross multiplication helper
}
\author{
Kai Guo
}
\keyword{internal}
