% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{cross_sum}
\alias{cross_sum}
\title{Cross sum helper}
\usage{
cross_sum(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Matrix of pairwise sums
}
\description{
Cross sum helper
}
\author{
Kai Guo
}
\keyword{internal}
