% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_with.R
\name{df2annolist}
\alias{df2annolist}
\title{Convert data frame to annotation list}
\usage{
df2annolist(df)
}
\arguments{
\item{df}{Data frame with Term and GeneID columns}
}
\value{
Named list of gene vectors
}
\description{
Convert data frame to annotation list
}
\author{
Kai Guo
}
