% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getdb}
\alias{.getdb}
\title{Get Database Package Name for a Given Species}
\usage{
.getdb(species = "human")
}
\arguments{
\item{species}{A character string specifying the species name (e.g., "human", "mouse").}
}
\value{
A character string containing the database package name corresponding to the input species.
}
\description{
This function returns the appropriate organism database package name for the specified species.
The package name is used for ID conversion (e.g. via AnnotationDbi) and is determined by a lookup
table. Allowed species are: "anopheles", "arabidopsis", "bovine", "celegans", "canine", "fly",
"zebrafish", "ecoli", "ecsakai", "chicken", "human", "mouse", "rhesus", "malaria", "chipm",
"rat", "toxoplasma", "streptomyces", "pig", "yeast", "xenopus", and "warm". If an unsupported
species is provided, the function returns "unsupported".
}
\details{
For example, for "human" the function returns "org.Hs.eg.db". This function is intended for
internal use in the package.
}
\examples{
.getdb("human")
.getdb("mouse")

}
\keyword{internal}
