% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getspeices}
\alias{.getspeices}
\title{Get Abbreviated Species Code}
\usage{
.getspeices(species = "human")
}
\arguments{
\item{species}{A character string specifying the species name (e.g., "human", "arabidopsis").}
}
\value{
A character string containing the abbreviated species code.
}
\description{
This function returns the abbreviated species code for the specified species.
The abbreviated code is used in KEGG queries and internal references.
}
\details{
For example, "human" is mapped to "hsa", "mouse" to "mmu", and "fly" to "dme". This
function uses a lookup table and is intended for internal use in the package.
}
\examples{
.getspeices("human")
.getspeices("arabidopsis")

}
\keyword{internal}
