% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKEGG.R
\name{.idc}
\alias{.idc}
\title{Convert IDs using an organism database}
\usage{
.idc(species, keys, fkeytype, tkeytype)
}
\arguments{
\item{species}{A string specifying the species code (e.g., "human","mouse","rat").}

\item{keys}{A vector of IDs to convert.}

\item{fkeytype}{The key type of the input IDs.}

\item{tkeytype}{The target key type (e.g., "SYMBOL").}
}
\value{
A vector of converted IDs.
}
\description{
This helper function converts a vector of keys from one type to another using AnnotationDbi::mapIds.
}
\keyword{internal}
