% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_genes_from_terms}
\alias{extract_genes_from_terms}
\title{extract genes based on index}
\usage{
extract_genes_from_terms(tree, term_ids)
}
\value{
vector of gene id
}
\description{
extract genes based on index
}
\author{
Kai Guo
}
