% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{filterByEvidence}
\alias{filterByEvidence}
\title{Filter annotation table by evidence codes}
\usage{
filterByEvidence(tb, evidenceCode)
}
\description{
Filter annotation table by evidence codes
}
\author{
Kai Guo
}
\keyword{internal}
