% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findMaxAncestor}
\alias{findMaxAncestor}
\title{Find maximum value ancestor between term pairs}
\usage{
findMaxAncestor(
  tree,
  terms,
  values,
  returnNames = FALSE,
  distanceMethod = c("longest", "shortest"),
  verbose = TRUE
)
}
\arguments{
\item{tree}{OntologyTree object}

\item{terms}{Character vector or numeric vector of term IDs}

\item{values}{Numeric vector of values for each term}

\item{returnNames}{Whether to return term names instead of IDs}

\item{distanceMethod}{Distance method ("longest" or "shortest")}

\item{verbose}{Whether to print progress messages}
}
\value{
Matrix of ancestor IDs or names
}
\description{
Find maximum value ancestor between term pairs
}
\author{
Kai Guo
}
