% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findMICA}
\alias{findMICA}
\title{Calculate Most Informative Common Ancestor (MICA) values}
\usage{
findMICA(tree, terms, method = "annotation", useCache = TRUE, verbose = TRUE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{terms}{Character vector or numeric vector of term IDs}

\item{method}{Method for calculating Information Content}

\item{useCache}{Whether to use cached values}

\item{verbose}{Whether to print progress messages}
}
\value{
Matrix of MICA values
}
\description{
Calculate Most Informative Common Ancestor (MICA) values
}
\author{
Kai Guo
}
