% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findMICAterm}
\alias{findMICAterm}
\title{Find Most Informative Common Ancestor (MICA) between terms}
\usage{
findMICAterm(
  tree,
  terms,
  method = "annotation",
  useCache = TRUE,
  returnNames = TRUE,
  distanceMethod = c("longest", "shortest"),
  verbose = TRUE
)
}
\arguments{
\item{tree}{OntologyTree object}

\item{terms}{Character vector or numeric vector of term IDs}

\item{method}{Method for calculating Information Content}

\item{useCache}{Whether to use cached values}

\item{returnNames}{Whether to return term names instead of IDs}

\item{distanceMethod}{Distance method ("longest" or "shortest")}

\item{verbose}{Whether to print progress messages}
}
\value{
Matrix of MICA term IDs or names
}
\description{
Find Most Informative Common Ancestor (MICA) between terms
}
\author{
Kai Guo
}
