% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{fromGraphToOntology}
\alias{fromGraphToOntology}
\title{Create ontology tree from igraph object}
\usage{
fromGraphToOntology(graph, edgeTypes = NULL, verbose = TRUE)
}
\arguments{
\item{graph}{An igraph object}

\item{edgeTypes}{Optional vector of edge relation types}

\item{verbose}{Print progress messages}
}
\value{
OntologyTree object
}
\description{
Create ontology tree from igraph object
}
\author{
Kai Guo
}
