% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{getAnnotations}
\alias{getAnnotations}
\title{count the annotation numbers}
\usage{
getAnnotations(tree, terms = NULL, uniquify = TRUE, useCache = TRUE)
}
\description{
count the annotation numbers
}
\author{
Kai Guo
}
