% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{getAnnotationsdb}
\alias{getAnnotationsdb}
\title{Get and process annotations from organism database}
\usage{
getAnnotationsdb(
  orgDb,
  namespace = "BP",
  evidenceCode = NULL,
  keytype = "SYMBOL",
  verbose = TRUE
)
}
\arguments{
\item{orgDb}{Organism database object or name}

\item{namespace}{GO namespace ("BP", "MF", "CC")}

\item{evidenceCode}{Vector of evidence codes to include}

\item{keytype}{Type of gene IDs ("SYMBOL", "ENTREZID", "REFSEQ", "ENSEMBL")}

\item{verbose}{Print progress messages}
}
\description{
Get and process annotations from organism database
}
\keyword{internal}
