% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getChildren}
\alias{getChildren}
\title{Get children for a specific term}
\usage{
getChildren(tree, termId, asNames = TRUE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{termId}{Term ID to get children for}

\item{asNames}{Whether to return term names instead of indices}
}
\value{
Vector of child terms
}
\description{
Get children for a specific term
}
\author{
Kai Guo
}
