% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{getGORelationships}
\alias{getGORelationships}
\title{Get GO relationships from appropriate database}
\usage{
getGORelationships(namespace, verbose)
}
\description{
Get GO relationships from appropriate database
}
\keyword{internal}
