% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getOffspringCount}
\alias{getOffspringCount}
\title{Get number of offspring terms for each term}
\usage{
getOffspringCount(tree, terms = NULL, useCache = TRUE, includeSelf = FALSE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{terms}{Optional subset of terms to get counts for}

\item{useCache}{Whether to use cached counts}

\item{includeSelf}{Whether to include the term itself in counts}
}
\value{
Named integer vector of offspring counts
}
\description{
Get number of offspring terms for each term
}
\author{
Kai Guo
}
