% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getParents}
\alias{getParents}
\title{Get parents for a specific term}
\usage{
getParents(tree, termId, asNames = TRUE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{termId}{Term ID to get parents for}

\item{asNames}{Whether to return term names instead of indices}
}
\value{
Vector of parent terms
}
\description{
Get parents for a specific term
}
\author{
Kai Guo
}
