% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getTermStats}
\alias{getTermStats}
\title{Get term statistics from OntologyTree}
\usage{
getTermStats(tree, name)
}
\arguments{
\item{tree}{OntologyTree object}

\item{name}{Name of the statistic to retrieve}
}
\value{
Requested statistic data
}
\description{
Get term statistics from OntologyTree
}
\author{
Kai Guo
}
