% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leaf_df}
\alias{leaf_df}
\title{KEGG Leaf Hierarchy Data}
\format{
A data frame with 363 observations of 6 variables.
}
\source{
KEGG BRITE database.
}
\usage{
leaf_df
}
\description{
A data frame containing leaf-level hierarchical information extracted from the KEGG BRITE database.
}
\details{
This dataset provides the hierarchical classification of KEGG BRITE terms at the leaf level,
including the codes and annotations for three levels:
\describe{
  \item{Level1}{A character vector of Level1 KEGG codes.}
  \item{Level1Anno}{A character vector of Level1 annotations.}
  \item{Level2}{A character vector of Level2 KEGG codes.}
  \item{Level2Anno}{A character vector of Level2 annotations.}
  \item{Level3}{A character vector of Level3 KEGG codes.}
  \item{Level3Anno}{A character vector of Level3 annotations.}
}
}
\examples{
data("leaf_df")
str(leaf_df, 2)
}
\keyword{datasets}
