% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{listTermStats}
\alias{listTermStats}
\title{List available term statistics}
\usage{
listTermStats(tree)
}
\arguments{
\item{tree}{OntologyTree object}
}
\value{
Character vector of available statistic names
}
\description{
List available term statistics
}
\author{
Kai Guo
}
