% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeRelations}
\alias{normalizeRelations}
\title{Normalize relation types}
\usage{
normalizeRelations(relations)
}
\arguments{
\item{relations}{Character vector of relation types}
}
\value{
Normalized relation types
}
\description{
Normalize relation types
}
\author{
Kai Guo
}
\keyword{internal}
