% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OntologyTree-class}
\alias{OntologyTree-class}
\alias{OntologyTree}
\title{Ontology Tree Class Definition}
\description{
A class representing hierarchical relationships in ontology
}
\section{Slots}{

\describe{
\item{\code{termNames}}{Character vector of term names}

\item{\code{termCount}}{Integer scalar of total terms}

\item{\code{edgeCount}}{Integer scalar of total relations}

\item{\code{parentMap}}{List of parent term indices}

\item{\code{childMap}}{List of child term indices}

\item{\code{edgeTypes}}{List of semantic relations between terms}

\item{\code{relTree}}{Ontology tree object for relation types}

\item{\code{source}}{Character scalar marking data source}

\item{\code{rootNode}}{Integer scalar of root term index}

\item{\code{leafNodes}}{Integer vector of leaf term indices}

\item{\code{annotations}}{List containing term annotations}

\item{\code{n_parents}}{Vector containing the parents number}

\item{\code{n_children}}{Vector containing the children number}

\item{\code{metadata}}{Additional term metadata}

\item{\code{termStats}}{List for storing arbitrary term-level statistics}
}}

\author{
Kai Guo
}
