% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{parseEdgeList}
\alias{parseEdgeList}
\title{Parse edge list into parent and child components}
\usage{
parseEdgeList(edges)
}
\arguments{
\item{edges}{Character vector of edges}
}
\value{
List of parent and child vectors
}
\description{
Parse edge list into parent and child components
}
\author{
Kai Guo
}
\keyword{internal}
