% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{print.OntologyTree}
\alias{print.OntologyTree}
\title{Print an OntologyTree object}
\usage{
\method{print}{OntologyTree}(x, ...)
}
\arguments{
\item{x}{An OntologyTree object}

\item{...}{Additional arguments (ignored)}
}
\value{
None (prints summary to console)
}
\description{
Print an OntologyTree object
}
