% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/cluster_R.R
\name{print.term_clusters}
\alias{print.term_clusters}
\title{Print method for term_clusters objects}
\usage{
\method{print}{term_clusters}(x, ...)

\method{print}{term_clusters}(x, ...)
}
\description{
Print method for term_clusters objects

Print method for term_clusters objects
}
\author{
Kai Guo
}
