% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{print.TermClusters}
\alias{print.TermClusters}
\title{Print method for TermClusters objects}
\usage{
\method{print}{TermClusters}(x, ...)
}
\description{
Print method for TermClusters objects
}
\author{
Kai Guo
}
