% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{processAnnotationTable}
\alias{processAnnotationTable}
\title{Process annotation table to list format}
\usage{
processAnnotationTable(tb, id_col = "SYMBOL")

processAnnotationTable(tb, id_col = "SYMBOL")
}
\arguments{
\item{tb}{Annotation table with gene IDs and GO terms}

\item{id_col}{Column name for gene IDs}
}
\description{
Process annotation table to list format

Process annotation table to list format
}
\author{
Kai Guo
}
\keyword{internal}
