% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{removeTermStats}
\alias{removeTermStats}
\title{Remove term statistics from OntologyTree}
\usage{
removeTermStats(tree, name)
}
\arguments{
\item{tree}{OntologyTree object}

\item{name}{Name of the statistic to remove}
}
\value{
Updated OntologyTree object
}
\description{
Remove term statistics from OntologyTree
}
\author{
Kai Guo
}
