% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{reportMultipleRoots}
\alias{reportMultipleRoots}
\title{Report multiple root nodes found}
\usage{
reportMultipleRoots(roots, terms)
}
\arguments{
\item{roots}{Root indices}

\item{terms}{Term names}
}
\description{
Report multiple root nodes found
}
\author{
Kai Guo
}
\keyword{internal}
