% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,OntologyTree-method}
\alias{show,OntologyTree-method}
\title{Show method for OntologyTree}
\usage{
\S4method{show}{OntologyTree}(object)
}
\arguments{
\item{object}{An OntologyTree object}
}
\description{
Show method for OntologyTree
}
