% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{summary,OntologyTree-method}
\alias{summary,OntologyTree-method}
\title{Summary method for OntologyTree}
\usage{
\S4method{summary}{OntologyTree}(object)
}
\arguments{
\item{object}{An OntologyTree object}
}
\description{
Summary method for OntologyTree
}
