% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{summary.OntologyTree}
\alias{summary.OntologyTree}
\title{Summary of an OntologyTree object}
\usage{
summary.OntologyTree(object, ...)
}
\arguments{
\item{object}{An OntologyTree object}

\item{...}{Additional arguments (ignored)}
}
\value{
A structured summary of the ontology tree
}
\description{
Summary of an OntologyTree object
}
