% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/cluster_R.R
\name{summary.TermClusters}
\alias{summary.TermClusters}
\title{Summary method for TermClusters objects}
\usage{
summary.TermClusters(object, ...)

summary.TermClusters(object, ...)
}
\description{
Summary method for TermClusters objects

Summary method for TermClusters objects
}
