% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{term2id}
\alias{term2id}
\title{Convert term names to term IDs}
\usage{
term2id(tree, termNames, strict = TRUE, addNames = FALSE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{termNames}{Character vector of term names or numeric vector of term IDs}

\item{strict}{Whether to error on missing terms}

\item{addNames}{Whether to add term names to returned IDs}
}
\value{
Integer vector of term IDs
}
\description{
Convert term names to term IDs
}
\author{
Kai Guo
}
