% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{terms}
\alias{terms}
\alias{nTerms}
\alias{nEdges}
\alias{nLeaves}
\title{Get terms in an ontology tree}
\usage{
terms(tree)

nTerms(tree)

nEdges(tree)

nLeaves(tree)
}
\arguments{
\item{tree}{An OntologyTree object}
}
\value{
Various information about terms in the tree:
  * `terms()`: character vector of all term names
  * `nTerms()`: integer scalar of total term count
  * `nEdges()`: integer scalar of total edge count
  * `nLeaves()`: integer scalar of total leaf term count
}
\description{
Get terms in an ontology tree
}
\examples{
\dontrun{
# Create a simple tree
parents <- c("a", "a", "b", "b", "c", "d")
children <- c("b", "c", "c", "d", "e", "f")
tree <- buildOntologyTree(parents, children)

# Get term information
terms(tree)
nTerms(tree)
}

}
\author{
Kai Guo
}
