% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_with.R
\name{updateAnnotationsR}
\alias{updateAnnotationsR}
\title{Update annotations for OntologyTree object}
\usage{
updateAnnotationsR(tree, new_annot, use_background = FALSE)
}
\arguments{
\item{tree}{OntologyTree object}

\item{new_annot}{Named list of genes per term or data frame}

\item{use_background}{Whether to use original annotations}
}
\value{
Updated OntologyTree object
}
\description{
Update annotations for OntologyTree object
}
