% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R, R/utils.R
\name{validateAnnotatedTerms}
\alias{validateAnnotatedTerms}
\title{Validate and filter terms with zero annotations}
\usage{
validateAnnotatedTerms(tree, termIds)

validateAnnotatedTerms(tree, termIds)
}
\arguments{
\item{tree}{OntologyTree object}

\item{termIds}{Term IDs to validate}
}
\value{
A logical vector indicating which terms have annotations

Logical vector indicating which terms to keep
}
\description{
Validate and filter terms with zero annotations

Validate and filter terms with zero annotations
}
\author{
Kai Guo
}
\keyword{internal}
