% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{validateInputs}
\alias{validateInputs}
\title{Validate parent and child inputs}
\usage{
validateInputs(parents, children)
}
\arguments{
\item{parents}{Parent term vector}

\item{children}{Child term vector}
}
\description{
Validate parent and child inputs
}
\author{
Kai Guo
}
\keyword{internal}
