% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOTree.R
\name{validateInputsGO}
\alias{validateInputsGO}
\title{Validate function inputs}
\usage{
validateInputsGO(namespace, relations, orgDb, evidenceCode)
}
\description{
Validate function inputs
}
\author{
Kai Guo
}
\keyword{internal}
