% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/cluster_R.R
\name{print.summary.TermClusters}
\alias{print.summary.TermClusters}
\title{Print method for TermClusters summary}
\usage{
\method{print}{summary.TermClusters}(x, ...)

\method{print}{summary.TermClusters}(x, ...)
}
\description{
Print method for TermClusters summary

Print method for TermClusters summary
}
\author{
Kai Guo
}
